Add shortcode param to skip group check
When the FAQ shortcode is processed, it checks for the existence of any
FAQ groups, if a user wants to shortcut that check and skip directly to
a full FAQ output (e.g. they want a list of recently added FAQs,
etc...), they can use the `skip_group` param and set that to `true`.

Fixes #21